requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
pydantic-settings==2.2.1
pymongo==4.6.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
fastapi-mail==1.4.1